+++ src/main/java/com/toy4/domain/dutyHistory/controller/DutyHistoryAdminController.java
@@ -1,6 +1,6 @@
 package com.toy4.domain.dutyHistory.controller;
 
-import com.toy4.domain.dutyHistory.dto.ApprovedDutyResponse;
+import com.toy4.domain.dutyHistory.dto.response.ApprovedDutyResponse;
 import com.toy4.domain.dutyHistory.dto.request.DutyStatusUpdateRequest;
 import com.toy4.domain.dutyHistory.service.DutyHistoryService;
 import com.toy4.global.response.dto.CommonResponse;
